Rename 'illegal spat transition' -> 'event state progression'
