+++ jpo-conflictmonitor/src/test/java/us/dot/its/jpo/conflictmonitor/monitor/algorithms/event_state_progression/EventStateProgressionPropertiesTest.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.conflictmonitor.monitor.algorithms.spat_transition;
+package us.dot.its.jpo.conflictmonitor.monitor.algorithms.event_state_progression;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Test;
@@ -9,7 +9,6 @@ import org.springframework.kafka.test.context.EmbeddedKafka;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
-import us.dot.its.jpo.conflictmonitor.monitor.algorithms.event_state_progression.EventStateProgressionParameters;
 import us.dot.its.jpo.conflictmonitor.monitor.models.event_state_progression.PhaseStateTransition;
 import us.dot.its.jpo.conflictmonitor.monitor.models.event_state_progression.PhaseStateTransitionList;
 
@@ -24,7 +23,7 @@ import static org.hamcrest.Matchers.*;
 @EmbeddedKafka(partitions = 1, brokerProperties = { "listeners=PLAINTEXT://localhost:9092", "port=9092"})
 @DirtiesContext
 @Slf4j
-public class SpatTransitionPropertiesTest {
+public class EventStateProgressionPropertiesTest {
 
     @Autowired
     private EventStateProgressionParameters params;
